Conflicts:
	build.gradle
	tapestry-core/src/main/java/org/apache/tapestry5/corelib/components/Alerts.java
	tapestry-core/src/main/java/org/apache/tapestry5/services/TapestryModule.java
	tapestry-core/src/main/java/org/apache/tapestry5/test/PageTester.java
	tapestry-core/src/main/resources/org/apache/tapestry5/t5-alerts.css
	tapestry-core/src/test/groovy/org/apache/tapestry5/integration/app1/AlertsTests.groovy
	tapestry-core/src/test/resources/org/apache/tapestry5/integration/app1/pages/AlertsDemo.tml
	tapestry-json/build.gradle
	tapestry-yuicompressor/build.gradle
	tapestry-yuicompressor/src/main/java/org/apache/tapestry5/internal/yuicompressor/JavaScriptResourceMinimizer.java

+++ tapestry-core/src/main/java/org/apache/tapestry5/internal/structure/Page.java
@@ -50,10 +50,11 @@ public interface Page extends PageLifecycleCallbackHub
     {
         /**
          * Time, in milliseconds, to construct the page. This includes time to construct components inside the page,
-         * as well as hooking everything together. You'll often see that the first page is expensive to construct,
+         * as well as hooking everything together, and includes the execution of {@link org.apache.tapestry5.internal.structure.Page#loaded()}.
+         * You'll often see that the first page is expensive to construct,
          * and later pages that use a similar mix of components are very cheap.
          */
-        public final long assemblyTime;
+        public final double assemblyTime;
 
         /**
          * The total number of components in the page, including the root component. This does not include the number of mixins.
@@ -66,7 +67,7 @@ public interface Page extends PageLifecycleCallbackHub
          */
         public final int weight;
 
-        public Stats(long assemblyTime, int componentCount, int weight)
+        public Stats(double assemblyTime, int componentCount, int weight)
         {
             this.assemblyTime = assemblyTime;
             this.componentCount = componentCount;
